app-ml/requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
catboost==1.2.8
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
ConfigArgParse==1.7.1
contourpy==1.3.0
cycler==0.12.1
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_daq==0.6.0
dash_renderer==1.8.3
databricks-sdk==0.46.0
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
dynaconf==3.2.11
et_xmlfile==2.0.0
evidently==0.4.36
Faker==37.6.0
fastapi==0.116.1
fastapi-cli==0.0.10
fastapi-cloud-cli==0.1.5
fastjsonschema==2.21.2
filelock==3.19.1
Flask==3.0.3
flask-cors==6.0.1
Flask-Login==0.6.3
fonttools==4.56.0
fsspec==2025.7.0
gcsfs>=2023.6.0
gevent==25.5.1
geventhttpclient==2.3.4
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
grpcio==1.74.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
kiwisolver==1.4.7
litestar==2.17.0
litestar-htmx==0.5.0
llvmlite==0.43.0
loguru
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mlflow==2.20.4
mlflow-skinny==2.20.4
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
multipart==1.3.0
mypy_extensions==1.1.0
narwhals==2.3.0
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
openpyxl==3.1.5
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-exporter-otlp-proto-grpc==1.31.0
opentelemetry-exporter-otlp-proto-http==1.31.0
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
optuna==4.5.0
packaging==24.2
pandas==2.2.3
pandera==0.26.1
patsy==1.0.1
pillow==11.1.0
platformdirs==4.4.0
plotly==5.24.1
polyfactory==2.22.2
protobuf==5.29.3
prometheus_client==0.23.1
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
pyodbc==5.2.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.1
#pywin32==2.3.1
#pywin32
PyYAML==6.0.2
pyzmq==27.0.2
referencing==0.36.2
regex==2025.8.29
requests==2.32.3
retrying==1.4.2
rich==14.1.0
rich-click==1.8.9
rpds-py==0.27.1
rsa==4.9
scikit-learn==1.5.2
scipy==1.15.2
setuptools==80.9.0
shap==0.48.0
shapash==2.7.10
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
starlette==0.47.3
statsmodels==0.14.5
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
traitlets==5.14.3
typer==0.17.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.11.0
urllib3==2.3.0
uvicorn==0.35.0
waitress==3.0.2
watchdog==6.0.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.6
wrapt==1.17.2
wsproto==1.2.0
XlsxWriter==3.2.2
zipp==3.21.0
zope.event==5.1.1
zope.interface==7.2
